Branding (#4)
* feat(build-web): implement branded web ui build

Modifies the clone script to grab the tardigrade theme when setting the -b switch. Updated the cleaner to delete the tardigrade repo. Updated the build-web script to copy the files when the tardigrade folder exists.

* feat(build-web): implement branded web ui build

Modifies the clone script to grab the tardigrade theme when setting the -b switch. Updated the cleaner to delete the tardigrade repo. Updated the build-web script to copy the files when the tardigrade folder exists.

* feat(build-web): add message when copying branded ui

Added an echo message so that the user knows that the branded uis are being copied. Added the tardigrade folder to .gitignore.